refactor(109579): move requisições e estados compartilhados para view
