+++ src/components/screens/LancamentoInicial/RelatorioFinanceiro/index.tsx
@@ -1,160 +1,37 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
+import { Link } from "react-router-dom";
 
-import { getLotesSimples } from "services/lote.service";
-import { getGrupoUnidadeEscolar } from "services/escola.service";
-import { getMesesAnosSolicitacoesMedicaoinicial } from "services/medicaoInicial/dashboard.service";
-import { getRelatoriosFinanceiros } from "services/medicaoInicial/relatorioFinanceiro.service";
-import {
-  FiltrosInterface,
-  RelatorioFinanceiroInterface,
-  RelatorioFinanceiroResponse,
-} from "interfaces/relatorio_financeiro.interface";
+import { FiltrosInterface } from "interfaces/relatorio_financeiro.interface";
 
 import { Spin } from "antd";
 import { Filtros } from "./components/Filtros/Index";
-import { toastError } from "components/Shareable/Toast/dialogs";
 import { Paginacao } from "components/Shareable/Paginacao";
 
 import { MESES } from "constants/shared";
 import { STATUS_RELATORIO_FINANCEIRO } from "../constants";
+import {
+  MEDICAO_INICIAL,
+  RELATORIO_FINANCEIRO,
+  RELATORIO_CONSOLIDADO,
+} from "configs/constants";
 import "./styles.scss";
-
-type SelectOption = {
-  uuid: string;
-  nome: string;
-};
-
-type MultiSelectOption = {
-  value: string;
-  label: string;
-};
+import useView from "./view";
 
 export function RelatorioFinanceiro() {
-  const [lotes, setLotes] = useState<MultiSelectOption[]>([]);
-  const [gruposUnidadeEscolar, setGruposUnidadeEscolar] = useState<
-    SelectOption[]
-  >([]);
-  const [mesesAnos, setMesesAnos] = useState<SelectOption[]>([]);
-  const [carregando, setCarregando] = useState(true);
   const [filtros, setFiltros] = useState<FiltrosInterface>({});
-  const [relatoriosFinanceiros, setRelatoriosFinanceiros] = useState<
-    RelatorioFinanceiroInterface[]
-  >([]);
-  const [relatoriosFinanceirosResponse, setResponseEmpenhosResponse] =
-    useState<RelatorioFinanceiroResponse>();
-  const [paginaAtual, setPaginaAtual] = useState(1);
 
-  const getLotesAsync = async () => {
-    try {
-      const { data } = await getLotesSimples();
-      const lotesOrdenados = data.results.sort((loteA, loteB) => {
-        return loteA.diretoria_regional.nome < loteB.diretoria_regional.nome;
-      });
-      const lotes = lotesOrdenados.map((lote) => {
-        return {
-          value: lote.uuid,
-          label: `${lote.nome} - ${lote.diretoria_regional.nome}`,
-        };
-      });
-      setLotes(lotes);
-    } catch (error) {
-      toastError("Erro ao carregar lotes. Tente novamente mais tarde.");
-    }
-  };
-
-  const getGruposUnidades = async () => {
-    try {
-      const { data } = await getGrupoUnidadeEscolar();
-      const tiposUnidades = data.results.map((grupo) => ({
-        uuid: grupo.uuid,
-        nome: grupo.tipos_unidades
-          ?.map((unidade) => unidade.iniciais)
-          .join(", "),
-      }));
-
-      setGruposUnidadeEscolar(
-        [
-          {
-            uuid: "",
-            nome: "Selecione o tipo de UE",
-          },
-        ].concat(tiposUnidades)
-      );
-    } catch (error) {
-      toastError(
-        "Erro ao carregar tipos de unidade escolar. Tente novamente mais tarde."
-      );
-    }
-  };
-
-  const getMesesAnosAsync = async () => {
-    try {
-      const { data } = await getMesesAnosSolicitacoesMedicaoinicial({
-        status: "MEDICAO_APROVADA_PELA_CODAE",
-      });
-      const mesesAnos = data.results.map((mesAno) => ({
-        uuid: `${mesAno.mes}_${mesAno.ano}`,
-        nome: `${MESES[parseInt(mesAno.mes) - 1]} de ${mesAno.ano}`,
-      }));
-      setMesesAnos(
-        [
-          {
-            uuid: "",
-            nome: "Selecione o mês de referência",
-          },
-        ].concat(mesesAnos)
-      );
-    } catch (error) {
-      toastError(
-        "Erro ao carregar meses de referência. Tente novamente mais tarde."
-      );
-    }
-  };
-
-  const getRelatoriosFinanceirosAsync = async (
-    page: number = null,
-    filtros: FiltrosInterface = null
-  ) => {
-    try {
-      filtros = { ...filtros, lote: filtros?.lote?.toString() };
-
-      const { data } = await getRelatoriosFinanceiros(page, filtros);
-
-      setRelatoriosFinanceiros(data.results);
-      setResponseEmpenhosResponse(data);
-    } catch (error) {
-      toastError(
-        "Erro ao carregar relatórios financeiros. Tente novamente mais tarde."
-      );
-    }
-  };
-
-  const requisicoesPreRender = async (): Promise<void> => {
-    Promise.all([
-      getLotesAsync(),
-      getGruposUnidades(),
-      getMesesAnosAsync(),
-      getRelatoriosFinanceirosAsync(),
-    ]).then(() => {
-      setCarregando(false);
-    });
-  };
-
-  useEffect(() => {
-    setPaginaAtual(1);
-    requisicoesPreRender();
-  }, []);
+  const view = useView({ filtros });
 
   const onChangePage = async (page: number, filtros: FiltrosInterface) => {
-    setPaginaAtual(page);
-    setCarregando(true);
-    await getRelatoriosFinanceirosAsync(page, filtros);
-    setCarregando(false);
+    view.setPaginaAtual(page);
+    view.setCarregando(true);
+    await view.getRelatoriosFinanceirosAsync(page, filtros);
+    view.setCarregando(false);
   };
 
   return (
     <div className="relatorio-financeiro">
-      <Spin tip="Carregando..." spinning={carregando}>
+      <Spin tip="Carregando..." spinning={view.carregando}>
         <div className="card mt-3">
           <div className="card-body">
             <Filtros
@@ -166,13 +43,13 @@ export function RelatorioFinanceiro() {
                 setFiltros({});
                 onChangePage(1, {});
               }}
-              lotes={lotes}
-              gruposUnidadeEscolar={gruposUnidadeEscolar}
-              mesesAnos={mesesAnos}
+              lotes={view.lotes}
+              gruposUnidadeEscolar={view.gruposUnidadeEscolar}
+              mesesAnos={view.mesesAnos}
             />
 
             <div className="mt-4">
-              {relatoriosFinanceiros.length === 0 && !carregando ? (
+              {view.relatoriosFinanceiros.length === 0 && !view.carregando ? (
                 <div className="text-center mt-4 mb-4">
                   Nenhum resultado encontrado
                 </div>
@@ -185,18 +62,19 @@ export function RelatorioFinanceiro() {
                   <table>
                     <thead>
                       <tr className="row">
-                        <th className="col-4">DRE</th>
+                        <th className="col-3">DRE</th>
                         <th className="col-3">Tipo de UE</th>
                         <th className="col-1 text-center">Lote</th>
                         <th className="col-2 text-center">Mês de Referência</th>
                         <th className="col-2 text-center">Status</th>
+                        <th className="col-1 text-center">Ações</th>
                       </tr>
                     </thead>
 
                     <tbody>
-                      {relatoriosFinanceiros.map((relatorioFinanceiro) => (
+                      {view.relatoriosFinanceiros.map((relatorioFinanceiro) => (
                         <tr key={relatorioFinanceiro.uuid} className="row">
-                          <td className="col-4">
+                          <td className="col-3">
                             {relatorioFinanceiro.lote.diretoria_regional.nome}
                           </td>
                           <td className="col-3">
@@ -217,6 +95,18 @@ export function RelatorioFinanceiro() {
                               ]
                             }
                           </td>
+                          <td className="col-1 text-center">
+                            <Link
+                              to={`/${MEDICAO_INICIAL}/${RELATORIO_FINANCEIRO}/${RELATORIO_CONSOLIDADO}/?uuid=${relatorioFinanceiro.uuid}`}
+                            >
+                              <span className="px-2">
+                                <i
+                                  title="Visualizar Relatório Consolidado"
+                                  className="fas fa-eye green"
+                                />
+                              </span>
+                            </Link>
+                          </td>
                         </tr>
                       ))}
                     </tbody>
@@ -224,9 +114,9 @@ export function RelatorioFinanceiro() {
 
                   <Paginacao
                     onChange={(page: number) => onChangePage(page, filtros)}
-                    total={relatoriosFinanceirosResponse?.count}
-                    pageSize={relatoriosFinanceirosResponse?.page_size}
-                    current={paginaAtual}
+                    total={view.relatoriosFinanceirosResponse?.count}
+                    pageSize={view.relatoriosFinanceirosResponse?.page_size}
+                    current={view.paginaAtual}
                   />
                 </div>
               )}
